+++ js/__tests__/fixtures/mock-cpanel-proxy.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  addAddonDomain: async () => { return { status: 1, errors: [] } },
+}